run healthchecks in background only
